requirements.txt
Django==1.6
djangorestframework==3.0.5
